{{ cookiecutter.project_slug }}/api/requirements.txt
django-ninja==0.21.0
ipdb
black==22.3.0
pytest
psycopg2-binary
snowflake-connector-python
sqlalchemy
celery

# Dev requirements
jupyterlab
ipython
django-extensions
